requirements.txt
python-pptx
pytz
requests
selenium
splinter
urllib3